chnged name of logs app and added .logs to gitignore
