--- va_explorer/logs/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class LogsConfig(AppConfig):
-    name = 'va_explorer.logs'
-    verbose_name = {'logs'}
