+++ va_explorer/va_logs/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class VaLogsConfig(AppConfig):
+    name = 'va_explorer.va_logs'
+    verbose_name = {'va_logs'}
